backend/requirements.txt
Flask==3.0.3
flask-cors==4.0.1
Flask-Limiter==3.8.0
gunicorn==22.0.0
python-dotenv==1.0.1
supabase==2.28.0
pytest==8.3.3
requests==2.32.3
boto3==1.37.0
cryptography==44.0.2
